ajout de dates

+++ tasks.php
@@ -20,9 +20,11 @@ function handlePost() {
     $repo = new TaskRepository(); 
     try {
         
-    $description = $_POST['description'];
-    $status = $_POST['status'];
-    $task = new Task(0, $description, $status);
+        $task = tryGetTask($_POST);
+        if (gettype($task) === "string") {
+            writeError($task);
+            return;
+        }
 
         $newTask = $repo->create($task);
         render($newTask);
@@ -31,13 +33,17 @@ function handlePost() {
         http_response_code(500);
         renderStatusMessage("Erreur à la création de la tâche : " . $e->getMessage(), MessageStatus::Error);
     }
-}
+}   
 
 
 function handleDelete() {
     parse_str(file_get_contents("php://input"), $body);
     try {
         $id = $body['id'];
+        if (is_nan($id)) {
+            writeError();
+            return;
+        }
         $repo = new TaskRepository();
         $repo->delete($id);
         renderStatusMessage("La tâche a été correctement supprimée", MessageStatus::Success);
@@ -50,16 +56,64 @@ function handleDelete() {
 function handlePatch() {
     parse_str(file_get_contents("php://input"), $body);
     try {
-        $id = $body['id'];
-        $description = $body['description'];
-        $status = $body['status'];
-        $taskToUpdate = new Task($id, $description, $status);
+        $task = tryGetTask($body);
+        if (getType($task) == "string" || is_nan($body["id"]))  {
+            writeError($task);
+            return;
+        }
         $repo = new TaskRepository();
-        $repo->update($taskToUpdate);
-        render($taskToUpdate);
+        $repo->update($task);
+        render($task);
         renderStatusMessage("La tâche a été correctement modifiée", MessageStatus::Success);
     } catch (Exception $e) {
         http_response_code(500);
         renderStatusMessage("Erreur à la création de la tâche : " . $e->getMessage(), MessageStatus::Error);
     }
+}
+
+/**
+ * Try to create a task from the body and returns null if it's not possible due to a bad request
+ */
+function tryGetTask(array $body): Task|string {
+
+    $keys = ["description", "status", "dueDate", "beginDate"];
+    foreach ($keys as $key) {
+        if (!array_key_exists($key, $body)) {
+            return "Informations manquantes : " . $key;
+        }
+    }
+
+    $desc = $body['description'];
+    $status = $body['status'];
+    $dueDate = $body["dueDate"];
+    $beginDate = $body["beginDate"];
+    if (!isValid($desc) || !isValid($status)) {
+        return "Données invalides";
+    }
+
+    if (isset($dueDate) && isset($beginDate) && $beginDate > $dueDate) {
+        return "La date de début ne peut pas être plus grande que la date de fin";
+    }
+
+    // check if the status is a valid TaskStatus
+    if (TaskStatus::tryFrom($status) === null) {
+        return "mauvais type de statut";
+    }
+
+    return new Task($body['id'] ?? 0, $desc, $dueDate, $beginDate, $status);
+}
+
+
+function isValidDate(string $date) {
+    return !isset($date) ? true :  strtotime($date) !== false;
+}
+
+function isValid(string $str) {
+    return isset($str) && strlen($str) > 0;
+}
+
+function writeError(string $message = "Entrées incorrectes") {
+    header('HX-Retarget: #message');
+    http_response_code(400);
+    renderStatusMessage($message, MessageStatus::Error);
 }
